+++ app/controllers/sessions_controller.rb
@@ -1,0 +1,7 @@
+class SessionsController < Devise::SessionsController
+
+  respond_to :json
+
+  skip_before_action :verify_authenticity_token, only: [ :create ]
+
+end